Build Linux libc++ configs (#35)
Add Linux libc++ configs that relate to the baremetal libc++ configs
we are building today.

For better or worse, these roughly match what we were doing in the past
with only minor cleanup. There's still lots of weirdness here (relying
on libgcc and ld, adding the GNU sysroot to source Linux headers),
but we're ignoring that for now with the aim of a) returning to those
problems soon and b) keeping this as a relatively small/isolated patch.

Signed-off-by: Jonathon Penix <jpenix@quicinc.com>