[SPIRV] Add support for non-interposable function aliases (#172730)
The backend was not handling GlobalAliases such as in call targets. This
patch pre-processes the aliases in the module and resolve them to their
aliasee when possible. The patch also documents those cases that are not
yet supported.

(cherry picked from commit 5b5237681ec10477721d28ddea957a1dc72dd42a)