AMDGPU: Stop requiring afn for f32 rsq formation (#172082)
We were checking for afn or !fpmath attached to the sqrt. We
are not trying to replace a correctly rounded rsqrt; we're replacing
the two correctly rounded operations with the contracted operation.
It's net a better precision, so contract on both instructions should
be sufficient. Both the contracted and uncontracted sequences pass
the OpenCL conformance test, with a lower maximum error contracted.

(cherry picked from commit 7d7d063c9c393ee936ebfe899cc673df18189e06)