[mlir][Interfaces] Add `RegionBranchOpInterface::getSuccessorOperands` helper (#173971)
Add a helper for querying the successor operands for a region branch
`src -> dst`. Both `src` and `dst` may be the region branch op itself or
a terminator.

This helper allows users to query successor operands for the region
branch op and the terminators in a uniform way. This is similar to
`getSuccessorRegions(RegionBranchPoint)`, which works both for region
branch ops and terminators.

(cherry picked from commit 60606b22a77b667afff8e08f2be61afcb3a3c379)