[Clang] Add NUW to the Sub in __builtin_clrsb expansion. (#174010)
The ctlz will produce a value in the range [1..bitwidth]. It can't
produce 0. This means the subtract of 1 will not have unsigned wrap.

It also has no signed wrap, but the optimizer can figure that out on its
own.

It's very likely InstCombine will just drop the NUW when it
canonicalizes to Add, but maybe it will be helpful in some case.

(cherry picked from commit 0bc6491da0c827ce94bd8fa852422eaafe6c39e7)