[mlir][Pass] Fix textual pipeline specification when generating a crashless reproducer (#173750)
Crashless reproducers currently emit an extra anchor op in the pipeline
they print out, because one gets added by `appendReproducer`. Fix this
by always adding the anchor op in the caller of `appendReproducer`.

In addition, `mlir-opt` always uses `any` as the anchor op, instead of
getting the anchor op from the `PassManager`, fix that as well so we can
test that the pipeline is reproduced as it was passed in.

(cherry picked from commit cb6c195de75a7a532129173df44a0ee8dda00a62)